+++ source/app/utilities/system/get_commands.py
@@ -12,9 +12,50 @@ def get_commands   ( commands ):
 	arguments = parse_commands ( commands )
 
 
+	if arguments [ 'help_menu' ]:
+
+		menu = [
+			'PlantUML class generator for JavaScript\n\n'
+			'python3 BuildClass.py {<source>} [<destination>] [flags] [args[|args...]]\n\n'
+			'PATHS . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .\n\n'
+			'source\t\t\t\tFile or directory location of javascript file(s) to convert\n\n'
+			'\t\t\t\tusage: \n'
+			'\t\t\t\t\t(single)    "/javascript/classes/one.js"\n'
+			'\t\t\t\t\t(multiple)  "/javascript/classes"\n\n'
+			'destination\t\t\tFile or directory location to save class diagrams\n\n'
+			'\t\t\t\tusage: \n'
+			'\t\t\t\t\t(single)    "/javascript/classes/output/one.txt"\n'
+			'\t\t\t\t\t(multiple)  "/javascript/classes/output"\n\n'
+			'FLAGS . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .\n\n'
+			'-o, --omit "<filename>"\t\tOmit the following filenames from the source directory\n\n'
+			'\t\t\t\tusage: \n'
+			'\t\t\t\t\t(single)   	--omit "file1"\n'
+			'\t\t\t\t\t(multiple) 	--omit "file1|file2|file3"\n\n'
+			'-s, --skin "<skinparam>"\tEmbed skin parameters within the class uml generated\n\n'
+			'\t\t\t\tusage: \n'
+			'\t\t\t\t\t(single)     --skin "skinparam+one+1"\n'
+			'\t\t\t\t\t(multiple)   --skin "skinparam+one+1|skinparam+two+2"\n\n'
+			'-m, --make "<image_type>"\tMake the class generated diagram into an image\n\n'
+			'\t\t\t\tusage: \n'
+			'\t\t\t\t\t(single) 	--make "png"\n'
+			'\t\t\t\t\t(multiple)	--make "png|svg|eps"\n\n'
+			'-l, --link\t\t\tLink available classes to generated class diagrams\n\n'
+			'\t\t\t\tusage: --link\n\n'
+			'-h, --help\t\t\tDisplay this help menu\n\n'
+			'\t\t\t\tusage: --help'
+		]
+
+		for line in menu:
+
+			print ( line )
+
+
+		return ERROR;
+
+
 	if arguments [ 'source' ] == None: 																# CHECK WHETHR A SOURCE IS PRESENT
 
-		print ( 'BuildClass.py requires a single source... try again !' )
+		print ( ' >> [ERROR] BuildClass.py\n\t~ Requires a single source !' )
 
 		return ERROR
 
@@ -23,7 +64,7 @@ def get_commands   ( commands ):
 
 		if is_directory ( arguments [ 'source'] ) and is_file ( arguments [ 'destination' ], None ):
 
-			print ( 'BuildClass.py a whole directory cannot be parsed into a single file... try again !' )
+			print ( ' >> [ERROR] BuildClass.py\n\t~ A whole directory cannot be parsed into a single file !' )
 
 			return ERROR