+++ source/app/utilities/system/list/create_2d_list.py
@@ -6,7 +6,7 @@ def create_2d_list ( depth ):
 
 	if depth < 1 or depth == None:
 
-		print ( 'create_2d_list.py requires a depth value param of at least 1 !' )
+		print ( ' >> [ERROR] create_2d_list.py\n\t~ Requires a depth value param of at least 1 !' )
 
 		return ERROR