+++ source/app/utilities/system/string/repeat_character.py
@@ -6,14 +6,14 @@ def repeat_character ( character, times ):
 
 	if len ( character ) != 1:
 
-		print ( 'repeat_character.py accepts a single character param only !' )
+		print ( ' >> [ERROR] repeat_character.py\n\t~ Accepts a single character param only !' )
 
 		return ERROR
 
 
 	if times < 2 or times == None:
 
-		print ( 'repeat_character.py accepts a times value of above 1 only !')
+		print ( ' >> [ERROR] repeat_character.py\n\t~ Accepts a times value of above 1 only !' )
 
 		return ERROR